+++ crates/forge/src/running.rs
@@ -7,7 +7,8 @@ use blockifier::execution::entry_point::{
 };
 use blockifier::execution::execution_utils::ReadOnlySegments;
 use blockifier::execution::syscalls::hint_processor::SyscallHintProcessor;
-use blockifier::state::cached_state::{CachedState, GlobalContractCache};
+use blockifier::state::cached_state::CachedState;
+use blockifier::state::errors::StateError;
 use blockifier::state::state_api::State;
 use cairo_felt::Felt252;
 use cairo_vm::serde::deserialize_program::HintParams;
@@ -18,13 +19,13 @@ use itertools::chain;
 use cairo_lang_casm::hints::Hint;
 use cairo_lang_casm::instructions::Instruction;
 use cairo_lang_runner::casm_run::hint_to_hint_params;
-use cairo_lang_runner::SierraCasmRunner;
 use cairo_lang_runner::{Arg, RunnerError};
+use cairo_lang_runner::{RunResult, SierraCasmRunner};
 use camino::Utf8Path;
 use cheatnet::constants as cheatnet_constants;
 use cheatnet::execution::contract_execution_syscall_handler::ContractExecutionSyscallHandler;
 use cheatnet::forking::state::ForkStateReader;
-use cheatnet::state::{CheatnetState, ExtendedStateReader};
+use cheatnet::state::{BlockInfoReader, CheatnetBlockInfo, CheatnetState, ExtendedStateReader};
 use conversions::StarknetConversions;
 use starknet::core::types::BlockTag::Latest;
 use starknet::core::types::{BlockId, BlockTag, MaybePendingBlockWithTxHashes};
@@ -38,6 +39,7 @@ use starknet_api::hash::StarkHash;
 use starknet_api::patricia_key;
 use starknet_api::transaction::Calldata;
 use test_collector::{ForkConfig, TestCase};
+use thiserror::Error;
 use tokio::runtime::Runtime;
 use tokio::sync::mpsc::Sender;
 use tokio::task::JoinHandle;
@@ -93,19 +95,21 @@ pub(crate) fn blocking_run_from_test(
         if send.is_closed() {
             return Ok(TestCaseSummary::Interrupted {});
         }
-        run_test_case(
-            args,
+        let run_result = run_test_case(
+            args.clone(),
             &case,
             &runner,
             &runner_config,
             &runner_params,
             &send_shut_down,
-        )
+        );
+
+        extract_test_case_summary(run_result, &case, args)
     })
 }
 
-fn build_context() -> EntryPointExecutionContext {
-    let block_context = cheatnet_constants::build_block_context();
+fn build_context(block_info: CheatnetBlockInfo) -> EntryPointExecutionContext {
+    let block_context = cheatnet_constants::build_block_context(block_info);
     let account_context = cheatnet_constants::build_transaction_context();
     EntryPointExecutionContext::new(
         block_context.clone(),
@@ -152,6 +156,16 @@ fn build_syscall_handler<'a>(
     Ok(syscall_handler)
 }
 
+#[derive(Debug, Error)]
+pub(crate) enum TestCaseRunError {
+    #[error("{0}")]
+    RunnerError(#[from] RunnerError),
+    #[error("{0}")]
+    StateError(#[from] StateError),
+    #[error("{0}")]
+    Error(#[from] anyhow::Error),
+}
+
 #[allow(clippy::too_many_arguments)]
 pub(crate) fn run_test_case(
     args: Vec<Felt252>,
@@ -160,7 +174,7 @@ pub(crate) fn run_test_case(
     runner_config: &Arc<RunnerConfig>,
     runner_params: &Arc<RunnerParams>,
     _send_shut_down: &Sender<()>,
-) -> Result<TestCaseSummary> {
+) -> Result<RunResult, TestCaseRunError> {
     let available_gas = if let Some(available_gas) = &case.available_gas {
         Some(*available_gas)
     } else {
@@ -168,7 +182,7 @@ pub(crate) fn run_test_case(
     };
     let func = runner.find_function(case.name.as_str())?;
     let initial_gas = runner.get_initial_available_gas(func, available_gas)?;
-    let runner_args: Vec<Arg> = args.clone().into_iter().map(Arg::Value).collect();
+    let runner_args: Vec<Arg> = args.into_iter().map(Arg::Value).collect();
 
     let (entry_code, builtins) = runner.create_entry_code(func, &runner_args, initial_gas)?;
     let footer = runner.create_code_footer();
@@ -179,7 +193,7 @@ pub(crate) fn run_test_case(
     );
     let (hints_dict, string_to_hint) = build_hints_dict(instructions.clone());
 
-    let state_reader = ExtendedStateReader {
+    let mut state_reader = ExtendedStateReader {
         dict_state_reader: cheatnet_constants::build_testing_state(
             &runner_params.predeployed_contracts,
         ),
@@ -189,9 +203,11 @@ pub(crate) fn run_test_case(
             &case.fork_config,
         )?,
     };
-    let mut context = build_context();
+    let block_info = state_reader.get_block_info()?;
+
+    let mut context = build_context(block_info);
     let mut execution_resources = ExecutionResources::default();
-    let mut blockifier_state = CachedState::new(state_reader, GlobalContractCache::default());
+    let mut blockifier_state = CachedState::from(state_reader);
     let syscall_handler = build_syscall_handler(
         &mut blockifier_state,
         &string_to_hint,
@@ -199,7 +215,10 @@ pub(crate) fn run_test_case(
         &mut context,
     )?;
 
-    let mut cheatnet_state = CheatnetState::default();
+    let mut cheatnet_state = CheatnetState {
+        block_info,
+        ..Default::default()
+    };
     let mut cheatable_syscall_handler =
         CheatableSyscallHandler::wrap(syscall_handler, &mut cheatnet_state);
     let contract_execution_syscall_handler =
@@ -215,26 +234,45 @@ pub(crate) fn run_test_case(
         &string_to_hint,
     );
 
-    match runner.run_function(
+    Ok(runner.run_function(
         runner.find_function(case.name.as_str())?,
         &mut test_execution_syscall_handler,
         hints_dict,
         instructions,
         builtins,
-    ) {
+    )?)
+}
+
+fn extract_test_case_summary(
+    run_result: Result<RunResult, TestCaseRunError>,
+    case: &TestCase,
+    args: Vec<Felt252>,
+) -> Result<TestCaseSummary> {
+    match run_result {
         Ok(result) => Ok(TestCaseSummary::from_run_result(result, case, args)),
 
         // CairoRunError comes from VirtualMachineError which may come from HintException that originates in the cheatcode processor
-        Err(RunnerError::CairoRunError(error)) => Ok(TestCaseSummary::Failed {
+        Err(TestCaseRunError::RunnerError(RunnerError::CairoRunError(error))) => {
+            Ok(TestCaseSummary::Failed {
+                name: case.name.clone(),
+                msg: Some(format!(
+                    "\n    {}\n",
+                    error.to_string().replace(" Custom Hint Error: ", "\n    ")
+                )),
+                arguments: args,
+                fuzzing_statistic: None,
+            })
+        }
+        // ForkStateReader.get_block_info() may return error
+        Err(TestCaseRunError::StateError(error)) => Ok(TestCaseSummary::Failed {
             name: case.name.clone(),
             msg: Some(format!(
                 "\n    {}\n",
-                error.to_string().replace(" Custom Hint Error: ", "\n    ")
+                error.to_string().replace(" : ", "\n    ")
             )),
             arguments: args,
             fuzzing_statistic: None,
         }),
-
         Err(err) => Err(err.into()),
     }
 }